some restructuring for better organization
