+++ commons/Rank.py
@@ -19,12 +19,20 @@ class Rank(Enum):
     def __lt__(self, other) -> bool:
         return self.value < other.value
 
+    def __repr__(self) -> str:
+        if self.value < 11:
+            return str(self.value)
+        return ['J', 'Q', 'K', 'A'][self.value - 11]
+
+    def __str__(self) -> str:
+        return self.__repr__()
+
     def get_value(self):
         return self.value
 
     def get_alphabet_representation(self) -> str:
         """
-        get_alphabet_representation 
+        get_alphabet_representation
         This is same as str for 2-9 but for 10, J, Q, K, A, returns those
         characters to print them on screen.
 
@@ -37,9 +45,3 @@ class Rank(Enum):
             return str(self.get_value())
 
         return ("T", "J", "Q", "K", "A")[self.get_value() - 10]
-
-
-if __name__ == '__main__':
-    for rank in Rank:
-        print(rank, end="\t")
-        print(rank.value)
